localmanus-backend/requirements.txt
fastapi
uvicorn
agentscope
pydantic
websockets
python-multipart
python-dotenv
duckduckgo-search
requests
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
